tGraphicDev, tvutil, vs23_gfx: fix passing of negative fill color
Fixes issue #12.

+++ libraries/TTBAS_LIB/tGraphicDev.h
@@ -26,8 +26,8 @@ class tGraphicDev {
     // グラフィック描画
     virtual void pset(int16_t x, int16_t y, uint8_t c);
     virtual void line(int16_t x1, int16_t y1, int16_t x2, int16_t y2, uint8_t c);
-    virtual void circle(int16_t x, int16_t y, int16_t r, uint8_t c, int8_t f);
-    virtual void rect(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t c, int8_t f);
+    virtual void circle(int16_t x, int16_t y, int16_t r, uint8_t c, int f);
+    virtual void rect(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t c, int f);
     //virtual void cscroll(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t d);
     virtual void gscroll(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t mode);
     virtual void set_gcursor(uint16_t, uint16_t);
